Adds minimum player count of 2 to GMM Station Spotlight Trait (#93740)
## About The Pull Request

What it says on the tin, the GMM Station Spotlight trait will not
generate a spotlight if there isn't more than 1 player on the station.

## Why It's Good For The Game

People have been running into the GMM Trait triggering during offline
testing, and functionally it makes sense to not apply the effect if
there's no-one around to potentially act on that information. If you're
alone, and the richest person on the station, that's functionally not
going to change any time soon.

## Changelog
:cl:
qol: The GMM Spotlight station trait will not create a lightbeam if
there's not more than 1 player alive.
/:cl:

---------

Co-authored-by: Zephyr <12817816+ZephyrTFA@users.noreply.github.com>