Adds a plasma stabilizers and thermal regulator to the Operative equipment locker. (#93843)
## About The Pull Request

What it says on the tin

## Why It's Good For The Game

These modules are often found mapped in roundstart, and are also
printable extremely early in the round. They're not exactly high power
modules, and they are largely important quality of life tools for
non-humans species.

Which, now, operatives can be. So if the op cares to get it, they can
shove it in their suit. The complexity limits of their suit will more
than likely require them to be picky about what they put in anyway.
That's why the complexity system exists in the first place.

They're best placed here so we can assume by default the operative
doesn't need them, since most species don't need, for example, plasma
stabilization. It might be important for icemoon in case an operative
has a nasty fall! ~~Or experiences that weird bug where plasmification
spreads to people~~

## Changelog
:cl:
qol: Operative lockers come with thermal regulator and plasma stabilizer
modules.
/:cl: