saftey checks for play_ambience (#93942)
## About The Pull Request
first part with the 1 minute return is so you can have a area without an
ambience without insane runtimes

second part: In the course of making
https://github.com/DarkPack13/SecondCity/pull/246 I experienced some
weird bugs where it was playing the sound but failing to get the length.
This notifies you why the fuck its happening cause its otherwise really
confusing
## Why It's Good For The Game
I spent a while debugging it, still REALLY confused why the bug is
actually happening in the first place, its somehow referencing a file
that's both gone, and no longer being referenced anywhere in the repo.
Either way this shit annoyed me lol.