Removes basetype penguin from the WABBAJACK_ANIMAL list (#94049)
## About The Pull Request

Title. The basetype penguin lacks an `icon_state` as well as a name.
This is the only use of the basetype penguin that could lead it to being
spawned, so I believe it's unintentional.

also adds it to `abstract_mob_types`.

## Why It's Good For The Game

Being turned into something spriteless is fairly confusing.

## Changelog
:cl:

fix: Several things that turn you into a random animal will no longer
turn you into a spriteless penguin.

/:cl: