Fix MULEbots not unloading cargo in the right direction (#94127)
## About The Pull Request

MULEbots and MULEbot beacons are set up so that the bot knows what
direction to push the crate in after delivery, so it can go into the
little delivery area and be seen by people. Unfortunately when it calls
`load.forceMove(loc)` in `unload()` that triggers `Exited()` which has a
condition for if the thing that Exited is `load` which calls `unload(0)`
and fucks it all up, dropping the crate on the tile with the plastic
flaps where nobody can see it. Changes the if branch to null `load`
earlier (we still have a reference as `cached_load`) so `Exited` won't
interfere

## Why It's Good For The Game

MULEbots are already a struggle and it doesn't help that nobody can see
the cargo it delivers cause it's on the wrong tile

## Changelog
:cl:
fix: fixed MULEbots unloading deliveries onto the wrong tile
/:cl: