No more bluespace inhalers (#94306)
## About The Pull Request

Fixes #94284

When someone walks away from you after you try to put an inhaler to
their mouth, it'll cancel the do_after. Currently the do_after is based
on the item in your hand and not the mob you're targeting, which leads
to that issue. If you're just targeting yourself the effect is the same,
so if the bar appearing on the item was a design intention I can revert
it if you're only targeting yourself.

Also, the function for using the inhaler returns ITEM_INTERACT_ stuff
when it should so you stop whacking yourself and others with the inhaler
after using it.

## Why It's Good For The Game

bugfix

## Changelog

:cl:
fix: inhalers no longer have infinite reach
fix: using an inhaler no longer results in whacking the target with it
afterwards
/:cl:

Co-authored-by: Fghj240 <fakeemail@notrealemail.com>