You can punch meteors (#94347)
## About The Pull Request

With legendary athletic ability or a loaded powerfist or a baseball bat
+ expert athletic ability, you can punch a meteor mid-flight to deflect
in in the **_general_** direction you hit it. There's an achievement for
doing it with just your hands.

## Why It's Good For The Game

No longer do we have to live in fear of these limp lumps of dust. Our
strongest boxer miners and ~sleepiest~ buffest assistants no longer need
to feel helpless as they are barraged, but can stand up and fight for
their beloved Nanotrasen even in the darkest of times.

More interaction couldn't hurt meteors. 

A way to actually directly benefit from emagging meteor shields

## Changelog
:cl:

add: Syndicate Powerfist users and our most athletic employees have
discovered that if you punch a meteor with enough force, it can be
redirected.

/:cl: