Makes Mjor The Creative and his underlings are Spaceproof. (#94317)
Makes mjor the boss of the paperwizard spaceproof and his underlings

## Why It's Good For The Game

Mjor is a very hard fight to begin with. but never fought in its
intended way and always end in the following two scenarios.

1 leading him out to space/let atmos drain out to kill him
2 your a changeling or have some other cheesy revive tactic to kill it.

Realisticly since this is being in space and your alone you are probaly
not gonna kill it unless you have somebody back you up and even then it
gives a run for your money

## Changelog
:cl: Ezel
balance:Mjor the Creative saw trough your crafty tactics and he and his
creations no longer takes damage from space
/:cl: